backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dropchat-backend"
version = "0.1.0"
description = "No-Login PDF Chat Agent Backend - RAG pipeline with FastAPI"
authors = [
    {name = "shadmau"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # AI/ML dependencies
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.6.2",
    "langgraph-cli[inmem]",
    "faiss-cpu>=1.7.4",
    "tavily-python",

    # PDF processing
    "PyMuPDF>=1.23.0",

    # Utility dependencies
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",

    # Data handling
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",  # For testing FastAPI
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["api", "agent"]

[project.scripts]
dropchat-server = "api.main:main"